day142/elasticsearch-log-integration/requirements.txt
elasticsearch==8.13.0
pika==1.3.2
fastapi==0.111.0
uvicorn==0.30.1
aiohttp==3.9.5
pydantic==2.7.1
python-dateutil==2.9.0
pytest==8.2.2
pytest-asyncio==0.23.7
requests==2.32.3
websockets==12.0
jinja2==3.1.4
structlog==24.1.0
colorama==0.4.6
pyyaml==6.0.1
